+++ user/delay.h
@@ -5,8 +5,8 @@
 
 
 void delay_ms(uint32_t _tickDelay);
-    uint64_t getTick(void);
-    uint64_t millis(void);
-		uint64_t getGlobalTime(void);
-		void setGlobalTime(uint64_t globalTime);
+    uint32_t getTick(void);
+    uint32_t millis(void);
+		uint32_t getGlobalTime(void);
+		void setGlobalTime(uint32_t globalTime);
 #endif
